Moved a lot of Tasks based formatting to ColourText and __init__.py
