+++ python/Ganga/GPIDev/Lib/Tasks/IUnit.py
@@ -2,7 +2,9 @@ from __future__ import absolute_import
 from __future__ import print_function
 from Ganga.GPIDev.Base import GangaObject
 from Ganga.GPIDev.Schema import Schema, Version, SimpleItem, ComponentItem, FileItem, GangaFileItem
-from .common import markup, overview_colours, logger
+from .common import logger
+from Ganga.Utility.ColourText import status_colours, overview_colours, ANSIMarkup
+markup = ANSIMarkup()
 import Ganga.GPI as GPI
 from Ganga.Core.exceptions import ApplicationConfigurationError
 from Ganga.GPIDev.Base.Proxy import stripProxy
