+++ python/Ganga/Utility/ColourText.py
@@ -186,7 +186,7 @@ class Foreground(object):
 
 colour_objects = {'fg': Foreground(), 'bg': Background(), 'fx': Effects()}
 
-
+# Some additional helper functions from Tasks package
 def getColour(name):
     """ Get a colour code from the symbolic name: fg = Foreground(), bg = Background(), fx = Effects()
     The name examples fg.red, fx.normal, bg.white
@@ -194,3 +194,34 @@ def getColour(name):
     """
     x, y = name.split('.')
     return getattr(colour_objects[x], y)
+
+cols = {"black": (0, 0), "red": (0, 1), "green": (0, 2), "orange": (0, 3), "blue": (0, 4), "magenta": (0, 5), "cyan": (0, 6), "lgray": (0, 7),
+        "dgray": (6, 0), "lred": (6, 1), "lgreen": (6, 2), "yellow": (6, 3), "lblue": (6, 4), "pink": (6, 5), "lcyan": (6, 6), "white": (6, 7)}
+
+
+def col(f, b):
+    return '\033[%i%i;%i%im' % (4 + cols[b][0], cols[b][1], 3 + cols[f][0], cols[f][1])
+
+
+def fgcol(f):
+    return '\033[%i%im' % (3 + cols[f][0], cols[f][1])
+# Status and overview colours
+status_colours = {
+    'new': "",
+    'running': fgcol("green"),
+    'completed': fgcol("blue"),
+    'pause': fgcol("cyan"),
+    'running/pause': fgcol("cyan"),
+}
+overview_colours = {
+    'ignored': "",
+    'hold': col("black", "lgray"),
+    'ready': col("lgreen", "lgray"),
+    'running': col("black", "green"),
+    'completed': col("white", "blue"),
+    'attempted': col("black", "yellow"),
+    'failed': col("black", "lred"),
+    'bad': col("red", "lcyan"),
+    'unknown': col("white", "magenta"),
+    'submitted': col("black", "lgreen"),
+}
